+++ src/renderer/components/SongsPage/songsPage.tsx
@@ -1,3 +1,7 @@
+/* eslint-disable no-console */
+/* eslint-disable promise/always-return */
+/* eslint-disable consistent-return */
+/* eslint-disable no-nested-ternary */
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable react/no-array-index-key */
 /* eslint-disable react/destructuring-assignment */
@@ -6,6 +10,7 @@
 /* eslint-disable promise/catch-or-return */
 /* eslint-disable import/prefer-default-export */
 import React from 'react';
+// import { logger } from 'main/logger';
 import { Song } from './song';
 import DefaultSongCover from '../../../../assets/images/song_cover_default.png';
 
@@ -27,13 +32,59 @@ interface SongsPageProp {
 export const SongsPage = (props: SongsPageProp) => {
   const songsData: AudioInfo[] = [];
   const [songData, setSongData] = React.useState(songsData);
+  const [sortingOrder, setSortingOrder] = React.useState(
+    'aToZ' as SongsPageSortTypes
+  );
+
+  React.useEffect(() => {
+    window.api
+      .checkForSongs()
+      .then((audioInfoArray) => {
+        if (audioInfoArray) return setSongData(audioInfoArray);
+      })
+      .catch((err) => console.log(err));
+  }, []);
+
+  React.useEffect(() => {
+    if (songData && songData.length > 0) {
+      let sortedSongData: AudioInfo[];
+      if (sortingOrder === 'aToZ')
+        sortedSongData = songData.sort((a, b) =>
+          a.title.replace(/\W/gi, '') > b.title.replace(/\W/gi, '')
+            ? 1
+            : a.title.replace(/\W/gi, '') < b.title.replace(/\W/gi, '')
+            ? -1
+            : 0
+        );
+      else if (sortingOrder === 'artistName')
+        sortedSongData = songData.sort((a, b) =>
+          a.artists.join(',') > b.artists.join(',')
+            ? 1
+            : a.artists.join(',') < b.artists.join(',')
+            ? -1
+            : 0
+        );
+      else if (sortingOrder === 'dateAdded')
+        sortedSongData = songData.sort((a, b) => {
+          if (a.modifiedDate && b.modifiedDate) {
+            return new Date(a.modifiedDate).getTime() <
+              new Date(b.modifiedDate).getTime()
+              ? 1
+              : -1;
+          }
+          return 0;
+        });
+      else sortedSongData = songData;
+      setSongData(sortedSongData);
+    }
+  }, [songData, sortingOrder]);
+
   const songs = songData.map((song) => {
     return (
       <Song
         key={song.songId}
         title={song.title}
         artworkPath={song.artworkPath || DefaultSongCover}
-        path={song.path}
         duration={song.duration}
         songId={song.songId}
         artists={song.artists}
@@ -48,19 +99,23 @@ export const SongsPage = (props: SongsPageProp) => {
     );
   });
 
-  React.useEffect(() => {
-    window.api.checkForSongs().then((audioData: AudioInfo[] | undefined) => {
-      if (!audioData) return undefined;
-      else {
-        setSongData(audioData);
-        return undefined;
-      }
-    });
-  }, []);
-
   return (
     <div className="main-container songs-list-container">
-      <div className="title-container">Songs</div>
+      <div className="title-container">
+        Songs
+        <select
+          name="sortingOrderDropdown"
+          id="sortingOrderDropdown"
+          value={sortingOrder}
+          onChange={(e) =>
+            setSortingOrder(e.currentTarget.value as SongsPageSortTypes)
+          }
+        >
+          <option value="aToZ">A to Z</option>
+          <option value="dateAdded">Date added</option>
+          <option value="artistName">Artist</option>
+        </select>
+      </div>
       <div className="songs-container">{songs}</div>
     </div>
   );